+++ scripts/format.sh
@@ -7,7 +7,7 @@ dune build @fmt --auto-promote
 
 echo Formatting ReScript code...
 files=$(find runtime tests -type f \( -name "*.res" -o -name "*.resi" \) ! -name "syntaxErrors*" ! -name "generated_mocha_test.res" ! -path "tests/syntax_tests*" ! -path "tests/analysis_tests/tests*" ! -path "*/node_modules/*")
-./cli/rescript-legacy.js format $files
+./cli/rescript.js format $files
 
 echo Formatting JS code...
 yarn format